+++ changelog.md
@@ -6,6 +6,9 @@ tristan.fleury@labalancelle.yo.fr
 **Site intranet de l'association La Balancelle**    
 
 ------------
+## [05/12/2019] [v0.0.2][task61]
+    - modification envoie mail lors du rappel permanence
+        
 ## [03/12/2019] [v0.0.2][bug66]
     - suppression de l'alert en retour ajax de modification utilisateur
     - optimisation envoi mail (préparation du mail puis boucle sur la liste